refactor: move section helpers to lib
