+++ app/actions.ts
@@ -3,32 +3,9 @@
 import type { Node, Edge } from "reactflow"
 import semver from "semver"
 import { formatEdgeLabel } from "../lib/formatEdgeLabel"
+import { getSectionForPackage } from "../lib/sections"
 
 const GITHUB_RAW_BASE_URL = "https://raw.githubusercontent.com"
-
-export const PACKAGE_SECTION_MAP: Record<string, string> = {
-  "circuit-json": "Specifications",
-  "@tscircuit/props": "Specifications",
-  "schematic-symbols": "Specifications",
-  "@tscircuit/footprinter": "Specifications",
-  "jscad-fiber": "Specifications",
-  "circuit-to-svg": "Core Utility",
-  "jscad-electronics": "Core Utility",
-  "@tscircuit/core": "Core",
-  "@tscircuit/schematic-viewer": "UI Packages",
-  "@tscircuit/pcb-viewer": "UI Packages",
-  "@tscircuit/3d-viewer": "UI Packages",
-  "@tscircuit/eval": "Packaged Bundles",
-  "@tscircuit/runframe": "Packaged Bundles",
-}
-
-export function getSectionForPackage(pkg: string, repo: string): string {
-  return (
-    PACKAGE_SECTION_MAP[pkg] ||
-    PACKAGE_SECTION_MAP[repo] ||
-    "Downstream"
-  )
-}
 
 export interface DisplayNodeData {
   label: string // package name for display
